refactor: group image preview modules into an image/ subdirectory
